fix multi-statement execution for oracle databases
The main trick is that the go oracle-db-driver used by sqlrepl isn't able
to execute multiple statements in the same go; but it is able to execute
one block that contains multiple statements. So if a range is given that
contains multiple statements, this fix will wrap the statements in
a BEGIN..END block.